made a few updates
