+++ db.json
@@ -143,7 +143,7 @@
         "id": 15,
         "name": "Sarova Hotel",
         "image": "https://www.kayak.com/rimg/himg/a9/0f/b2/expediav2-67349-c13dae-274967.jpg?width=1366&height=768&crop=true",
-        "Description": "Mombasa Beach Hotel is under the flagship of Kenya Safari Lodges & Hotels is located in the Kenya's North Coast.  ",
+        "Description": "Sarova Hotel is under the flagship of Kenya Safari Lodges & Hotels is located in the Kenya's capital Nairobi.  ",
         "Category": "Hotel",
         "Location": "Nairobi"
